reorganize setup
